storage: Extract processor-related functionality
Plus some Processor refactorings. Specifically:

- move maxDownloadRetries from Storage to the Processor
- unexport WorkerPool type and its methods/struct fields
- move Perform() from Job to workerPool type and unexport it
- NewProcessor additionally receives the http.Client and log.Logger
  arguments
- workerPool contains a pointer to its processor (so that it can use the
  processor's http client to perform requests)
- export Processor's storageDir, scanInterval, Client and Logger members

Also temporarily disabled tests until we're finished with the
refactorings.

With this change, most of the remaining methods in storage.go are
related solely the storage part (apart from some leftovers that will be
moved in a subsequent step).

Also the Processor's API is reduced to 2 methods: NewProcessor() and
processor.Start(). Furthermore, there's now no coupling between the
Processor and the Storage.