Decouple Storage from Notifier and Processor
Storage, APIServer, Processor, Notifier, Job and Aggregation are now
distinct types.

Job and Aggregation are the core entities and therefore have their
corresponding types. All the other components typically interact with
those core types.

Storage provides the persistence layer and is used by the other components
that need to interact with it, notably Notifier, APIServer and Processor.

Many tests are broken and will be updated in a subsequent patch.

This change sets the groundwork for making each component a separate
package. Since there is now minimal to zero coupling between the packages, it
is easier to test, change and maintain them.

The way I imagine it to be in the future, is this:

  - The packages: Notifier, Processor, APIServer, Storage, (Job,
  Aggregation)
  - Storage imports (Job, Aggregation)
  - Notifier, Processor, APIServer import Storage and (Job, Aggregation)

Closes T6078