storage: Garbage collect unused aggregation keys
Before deleting the aggregation key, we want to make sure that there
isn't any job with the same ID in the jobs queue.

The operation has to be executed atomically since a new job may be
added right before we delete the aggregation, leaving the job with no
corresponding aggregation.

Fixes T6662