Use a default limit aggr if aggr key doesn't exist
With this commit we refactor the GetAggregation function to return an
Aggregation with the expected ID and a default limit if the key is not
found in Redis.

This mechanism will make sure that we can associate every job with a
valid aggregation key.

For example, a job may be leftover from an interrupted previous run. As
a rogue download it will get injected back to the queue eventually. The
aggregation for this job doesn't exist anymore because it got deleted
by the RemoveAggregation function. In this case the GetAggregation
will return an aggregation with the default limit so that the job can
be processed normally.