p/validator: Handle empty responses gracefully
Apparently rakyll/magicmime panics when given empty byte slices with
"runtime error: index out of range" when trying to convert to a C
buffer.

Note that this exposed some other logical errors on the way we handle
incomplete HTTP response reads. In such a case we should allow the
reader's error to surface and completely skip the validation. We will
handle this in a separate PR.

Closes T8945