p/mimetype: return read errors first, and skip validation checks
Previously we swallowed ErrUnexpectedEOF errors comming straight from
the response Reader. In such cases we continued with the mime validation
of the partial read blobs, and returned validation errors as usual.

The problem is that any possible mime-type errors can potentially mask
the HTTP errors that precedeed and possibly caused them.

In this patch we try not to use io.ReadAtLeast() since it can mask our
ErrUnexpectedEOF errors. In fact, we don't want to read at least 1024
bytes, but *at most* 1024 bytes (if EOF is reached earlier). We do that
by using a bytes.Buffer.ReadFrom() with a LimitedReader locked to 1024
bytes.

See T8945