Generalize notifier and add kafka as a backend
This commit refactors the codebase of the downloader in order
to introduce the concept of backends considering the notifier
component of the service. Below we present the most notable
changes.

* Package configuration
  - Introduce a map `Backends` in order to be able to add backends along with
    configuration parameters.

* Package job
  - Move `CallbackInfo` struct to package job and rename to `Callback`.
    Also add two new fields `Delivered` and `DeliveryError` to capture
    state considering the failure of communicating results.

* Package backend
  - Introduce the concept of backend and define a set of methods which
    concrete implementations must follow. A backend is responsible
    for communicating jobs' results through a medium/protocol such as
    HTTP or Kafka.

* Package http_backend
  - Implement the backend interface and communicate jobs results through HTTP.

* Package kafka_backend
  - Implement the backend interface and communicate jobs results through Kafka.

* Package notifier
  - Refactor the package to support backends. Start method recieves a map
    which contains configurations for enabled backends. Each backend
    recieves its configuration and then starts. According to a job's
    callback type and destination the appropriate backend is chosen
    in order to communicate jobs results. Once the program recieves
    the shutdown signal the backends are being gracefully stopped.

* Package main
  - Pass the backends of the config file to the Notifier component.

It is important to note the following. Previously Notifier used only HTTP for
communicating jobs results and when a 201 response was recieved
then a file deletion of the respective job was scheduled. Unfortunately,
due to the introduction of Kafka this behavior could not be maintained and
has been dropped. This is because of Kafka's asynchronous nature, since we can
not guarantee the delivery of a message to a consumer directly but we can only
guarantee the delivery of a message to a Kafka cluster. Http on the other hand
is synchronous and a response 200 up to 300 meant that the client has recieved
the callback and the downloaded resource using the download URL.
So, it is left to the user of this service to decide for strategy of when files
should be deleted.