Delete downloaded files after a time interval
This commit introduces a time interval after which a file of
a job will be scheduled for deletion if the job's callback is successful.
This time interval is a positive number and is specified in minutes.

This commit sets a common ground for a file deletion policy in order
to handle both HTTP which is synchronous and Kafka which is
asynchronous. Previously, only HTTP was present as a notification
mechanism and a file deletion was scheduled immediately after a 201
response was received from a job callback POST request.