Configure Http Client Download timeout per job
Add a new API option, `download_timeout` to configure separately http client
timeout per job.

Default timeout has been removed, for the case that
- user gives a timeout bigger than the default: it isn't respected and
  the default timeout is dominant.

We decided to avoid this inconsistency by removing completely the timeout from the client.

Timeout will be applied only if it is send with the rest of Job's
configuration, with this new API option `download_timeout`. Otherwise
there will be no timeout for the download client.