Differentiate the initial try from the retries
Working on implementing the `max-retries` option [1] surfaced a logical
error implemented in the retry mechanism.

We treat the original request as a *retry*.

The commit changes the number of default max retries to 2 (from 3) and
we now check the download count against the real number of retries allowed.

The tests do not require any changes since the implemented behavior is the
same.

We had:

```
1 retry
2 retries
3 retries
```
Now we have:

```
try
1 retry
2 retries
```

[1] https://github.com/skroutz/downloader/pull/17