Update workflows to build docker image
Building and publishing a docker image is only supported for the master branch.
This can be trigerred automatically by pushing a tag on the master branch or
manually using `workflow_dispatch`.

Image is labeled and tagged automatically using the `latest`,`<short_sha>` and
`<semver>` tags.

Image is published under the `ghcr.io/skroutz/downloader` project.