Processor: Create a filestorage interface and an AWS S3 implementation.
Processor stores files in the filesystem. With this commit it becomes possible
to choose through the config if the files will be stored on a local filesystem
or on an AWS S3 bucket.

Additionally, storing a file to a job specified S3 bucket is supported
using the "s3_bucket" and "s3_region" JSON keys. The job submitter
need to have taken care of the downloader's IAM access to the bucket.

A filestorage sub-section is introduced under the processor config section.

AWS access is expected to be configured through env vars or an AWS config file.