Add expiration field on Aggregation
In order to enable multiple processor instances run on the same redis we need
to make sure that each aggregation is only processed by one processor. Moreover
we need to make sure that if a processor dies, all of the corresponding jobs
are requeued.

This is achieved by adding `ExpiresAt` field on Aggregation. It is a Unix
Timestamp and is set when a processor picks up an aggregation. Regularly in the
processor loop, this expiration is refreshed so as to indicate that the
aggregation is still being processed and the processor is still alive.

In the event that a processor dies while an Aggregation is being processed,
we need to wait the Aggregation expiration time and then a new processor will
pick up the aggregation and requeue only the jobs that are left in `InProgress`
state. This refactors the logic behind `collectRogueDownloads()` as it is now
automated.