aws_ecs: Introduce an AWS ECS task definition for downloader
In order to be able to run Docker containers in AWS ECS, we need to
define a task definition for it[0].

Currently, downloader processes expect a JSON configuration file upon
start. Since, passing configuration files to ECS requires some extra
effort, we chose to modify downloader in order to be able to get those
variables from the given environment. As a temporary workaround until we
implement that change, we chose to pass the JSON config file via an SSM
Parameter in a base64 format since there are no sensitive values on it
Then using an entryPoint definition param[1], we decode and write this
config to the desired location for downloader to find it.

For the time being, the SSM Parameter was set manually on AWS side.

[0]: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definitions.html
[1]: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definition_parameters.html