workflows: Create a GHA to deploy downloader to AWS ECS
This commit introduces a GHA workflow for deploying downloader to Amazon
ECS on AWS Fargate. In short, the GHA performs the following actions[0]:

- Containerize the downloader application
- Push the image to Amazon Elastic Container Registry (ECR)
- Deploy the image to Amazon Elastic Container Service (ECS)

In order to allow GHA workflow access AWS resources, we make use of
OIDC[1] so there's no need to store AWS credentials as long-lived GitHub
secrets. To that end, authorization to AWS is a two step process; we
first need to assume the GitHub Action role using OIDC[2] and then
assume the application deployment role prior to execute any actions on
AWS.

[0]: https://docs.github.com/en/actions/deployment/deploying-to-your-cloud-provider/deploying-to-amazon-elastic-container-service
[1]: https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/about-security-hardening-with-openid-connect
[2]: https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-amazon-web-services