workflows: Avoid building and pushing the same image on ECR
Our Amazon ECR repositories are immutable, which means that we prevent
image tags from being overwritten by subsequent image pushes using the
same tag. So we cannot re-trigger an action that already run, since the
image push step will fail.

However, there are several cases where we need to just re-run the last
workflow without pushing a new commit on the application repository; for
example after making an infrastructure-specific modification. In order
to handle that scenario, we create an extra step which checks if the
image already exists on Amazon ECR. If the image exists, we simply skip
the build and push step and normally continue with the application
deployment.