workflows: Add a GHA to deploy downloader to AWS production account
This commit takes advantage of the existing `aws_ecs` reusable workflow
in order to deploy downloader to the respective AWS production account.

At the moment, we proceeded with a separate task definition for the
production environment; even though the differences between the two task
definitions are negligible we created a separate file mostly to move
fast and for readability. In the future we could consider converting the
task definition to a template which will be shared among different
environments.

Finally, keep in mind that the account id of each AWS account is visible
through the respective SSM parameter; given that this is an internal
repository we'll live with this for a while, until we properly hide that
information as well.