aws_ecs: Preserve proper tags on ECS task definitions
Initially, ECS task definitions are deployed via Terraform where we also
create the correct tags for the resources. Since, we chose to override
the task definition via a GHA we should ensure that the proper resource
tagging is preserved.

This commit adds a basic set of tags to production and staging task
definitions which include the following tags:

- Name
- Environment
- OwnedBy
- ManagedBy
- DeployedFrom
- map-migrated: required in order to get AWS credits