docker-build: Add docker-build targets to Makefile
This commit adds the following targets to Makefile:

- docker-image-build

This target builds the docker image to the local registry. The image is
as thin as possible including only the go compiler/runtime as well as
some C shared libraries.

- docker-build

This target builds the current code and upon success creates the
`docker-build/downloader` executable.

- docker-clean

Removes the `docker-build/` directory and its contents

This targets serve as a convenience for building downloader without
typing a shit ton of docker commands.