workflows: Setup DR deployment GitHub Action
This commit setups a new GHA for deploying downloader to the DR region.
It follows the exact same behavior as the production and staging GHAs
and we only have to push the latest commit to the `dr` branch.

Since DR deployment is an emergency task which we want to deploy fast
and given that the production deployed commit will already have the test
suite run, we skip running the tests suite and deploy directly to the
DR region. Last, note the the DR infra should be in-place otherwise the
application deployment will fail.